requirements.txt
python-telegram-bot
sentence-transformers
faiss-cpu
openai
python-dotenv
numpy
PyMuPDF